pyproject.toml
[project]
name = "spotify-stats"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "faker-file[common]>=0.18.4",
    "flower>=2.0.1",
    "psycopg[binary]>=3.2.9",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.8",
]

[tool.ruff]
fix = true

exclude = [
    ".git",
    ".venv",
    "migrations",
    "__pycache__"
]

lint.select = [
    "E",
    "F",
    "I",
    "UP",
    "N",
    "B",
    "C4",
    "RUF"
]

lint.ignore = [
    "E501",
    "B008",
    "RUF012",
    "F403",
    "F405"
]

line-length = 88


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
